pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "zion"
version = "0.0.1"
description = "A CLI tool for managing virtual machines"
authors = [{ name = "Dawid Butrykowski", email = "dawid.butrykowski@cerberus.org.pl" }]
license = {  text = "MIT" }
readme = "README.md"
keywords = ["virtualization", "libvirt", "cli"]
classifiers = [
    "Framework :: Click",
    "Libraries :: Libvirt, Pytest",
    "Programming Language :: Python :: 3",
]
dependencies = [
  "click==8.1.7",
  "hydra-core==1.3",
  "libvirt-python==10.5.0"
]
requires-python = ">=3.6"

[project.scripts]
zion = "zion.cli:cli"

[tool.setuptools]
packages = ["zion", "zion.cli", "zion.hypervisor", "zion.domain"]
